pyproject.toml
[project]
name = "whoop-telegram-bot-ai"
version = "0.1.0"
description = "Research project: Telegram bot with WHOOP integration and AI-powered productivity coaching. Uses LLM for personalized daily coaching through 6 daily slots. Personalized messages based on physical metrics from WHOOP."
readme = "README.md"
requires-python = ">=3.12"
keywords = ["telegram-bot", "whoop-api", "health-tracking", "personal-assistant", "productivity", "fitness", "wellness", "ai-assistant", "llm", "self-improvement", "whoop-integration", "productivity-coaching", "research-project"]
authors = [
    {name = "Aleksandr Mordvinov"}
]
dependencies = [
    "aiogram>=3.13.0",
    "apscheduler>=3.10.4",
    "httpx>=0.27.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.5.0",
    "pyright>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = []

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"

